+++ packages/react-router-manage/CHANGELOG.md
@@ -1,5 +1,17 @@
 # react-router-manage
 
+## 1.1.2-beta.1
+
+### Patch Changes
+
+- modify 'getCurrentRoute' get params route
+
+## 1.1.2-beta.0
+
+### Patch Changes
+
+- When configuring sub routes, the path can be empty
+
 ## 1.1.0
 
 ### Patch Changes